Fix SQL type casting for exportar-ficha endpoint - VARCHAR to INTEGER conversion
🔧 Fixed SQL type casting errors:
- Added ::integer cast for g.raza_id = r.id joins
- Added ::integer cast for g.user_id comparison
- Resolves operator does not exist: character varying = integer error

+++ app/api/v1/gallos_con_pedigri.py
@@ -1013,12 +1013,12 @@ async def exportar_ficha_gallo(
                 m.id as madre_id, m.nombre as madre_nombre, m.codigo_identificacion as madre_codigo,
                 m.foto_principal_url as madre_foto, mr.nombre as madre_raza
             FROM gallos g
-            LEFT JOIN razas r ON g.raza_id = r.id
+            LEFT JOIN razas r ON g.raza_id::integer = r.id
             LEFT JOIN gallos p ON g.padre_id = p.id
-            LEFT JOIN razas pr ON p.raza_id = pr.id
+            LEFT JOIN razas pr ON p.raza_id::integer = pr.id
             LEFT JOIN gallos m ON g.madre_id = m.id
-            LEFT JOIN razas mr ON m.raza_id = mr.id
-            WHERE g.id = :gallo_id AND g.user_id = :user_id
+            LEFT JOIN razas mr ON m.raza_id::integer = mr.id
+            WHERE g.id = :gallo_id AND g.user_id = :user_id::integer
         """)
         
         gallo_result = db.execute(query_gallo, {
